Finished writing first core-simulation made compilable

+++ FluidSim/fsmath.h
@@ -15,11 +15,11 @@ namespace fsmath {
 	inline float NextPowerOf2(float x)
 	{
 		float exponent = std::ceil(std::log2f(x));
-		return pow(2, exponent);
+		return static_cast<float>(pow(2, exponent));
 	}
 
-	inline glm::uvec3 allNextPowerOf2(const glm::uvec3 &vec) {
-		return glm::uvec3(
+	inline glm::uvec3 allNextPowerOf2(const glm::vec3 &vec) {
+		return glm::vec3(
 			NextPowerOf2(vec.x),
 			NextPowerOf2(vec.y),
 			NextPowerOf2(vec.z)
@@ -62,17 +62,15 @@ namespace fsmath {
 
 	void ComputeJacobian(UniformGrid<glm::mat3x3> &jacobianGrid, UniformGrid<glm::vec3> &velocityGrid);
 
-	//compute a finite difference at a point in a velocity grid and store the result in a matrix in the jacobian grid
-	//for "non"boundary points a centered scheme is used, else either a forward or a backward one is used
-
 	inline void ComputeFiniteDifference(UniformGrid<glm::mat3x3> &jacobianGrid, UniformGrid<glm::vec3> &velocityGrid,
 		const size_t xIndex, const size_t yIndex, const size_t zIndex, const size_t offset, const glm::vec3 &doubleCellExtent,
 		const glm::uvec3 &pointsAmountMinus1, const size_t yOffsetDistance, const size_t zOffsetDistance)
 
 	{
+
 		glm::mat3x3 &currentMatrix = jacobianGrid.AtOffset(offset);
 
-		// d/dx
+
 		if (xIndex == 0) {
 			currentMatrix[0] = (velocityGrid.AtOffset(offset + 1) - velocityGrid.AtOffset(offset)) / velocityGrid.GetCellExtent();
 		}
@@ -83,7 +81,7 @@ namespace fsmath {
 			currentMatrix[0] = (velocityGrid.AtOffset(offset + 1) - velocityGrid.AtOffset(offset - 1)) / doubleCellExtent;
 		}
 
-		// d/dy
+
 		if (yIndex == 0) {
 			currentMatrix[1] = (velocityGrid.AtOffset(offset + yOffsetDistance) - velocityGrid.AtOffset(offset)) / velocityGrid.GetCellExtent();
 		}
@@ -94,7 +92,7 @@ namespace fsmath {
 			currentMatrix[1] = (velocityGrid.AtOffset(offset + yOffsetDistance) - velocityGrid.AtOffset(offset - yOffsetDistance)) / doubleCellExtent;
 		}
 
-		// d/dz
+
 		if (zIndex == 0) {
 			currentMatrix[2] = (velocityGrid.AtOffset(offset + zOffsetDistance) - velocityGrid.AtOffset(offset)) / velocityGrid.GetCellExtent();
 		}
@@ -104,5 +102,7 @@ namespace fsmath {
 		else {
 			currentMatrix[2] = (velocityGrid.AtOffset(offset + zOffsetDistance) - velocityGrid.AtOffset(offset - zOffsetDistance)) / doubleCellExtent;
 		}
+
 	}
+
 }
